Medication crud operations.

+++ MedicalAssistant.Application/MedicationNotifications/Queries/GetMedicationNotificationByWeekQuery.cs
@@ -3,5 +3,5 @@ using MedicalAssistant.Application.Dto;
 
 namespace MedicalAssistant.Application.MedicationNotifications.Queries;
 public sealed record GetMedicationNotificationByWeekQuery(
-	int Offset,
+	double Offset,
 	DateTime Date) : IRequest<IEnumerable<MedicationNotificationDto>>;
